Add files via upload
In this project, the deep Reinforcement Learning (RL) controller is created to mainly deal with nonlinear swing-up problems related to the cart-pole system modelled as a multi-body dynamical system. Additionally, the deep RL controller's sensitivity analysis is done when it is applied to the cart-pole swing-up problem. In order to achieve this, the influence of altering the system's physical characteristics as well as the existence of dry friction forces, are examined using the task's cumulative reward. To demonstrate that the neural network architecture used in this study has sufficient learning potential to handle the task with modifications as high as a 90% increase in the pendulum mass and a 100% increment in the cart mass, extreme limits for the modifications of the parameters are found.